Configure environment for deploy
Configure some files to be ready for deploy in cloud

Put into .gitignore file of some keys of configuration

Create a secret for production and configure file secrets.yml

Create a new file docker-compose-prod.yml with respect configuration
to deploy our app to the cloud

Change the name of our services from website to app to be equals
either files docker-compose and docker-compose-prod.

As a result, the application can be ready for deploy with configuration
about enviroment.