Fix bug about returns privates questions
Fix bug when request to show a specific question that not is a public
question. To do that the consult was change from find_by to
find_by_id_and_private include another param as a false.

As a result, even someone request with a id that private questions it's
not showing for him, only showing if that question it's a public